scripts/requirements.txt
pyyaml>=6.0
openpyxl>=3.1
markdown>=3.5
weasyprint>=60.0
